Removes conditionalFormattingRule grouping by address, adds test
