+++ EPPlusTest/EPPlusTest.csproj
@@ -96,7 +96,7 @@
   <ItemGroup>
     <Compile Include="Address.cs" />
     <Compile Include="Calculation.cs" />
-    <Compile Include="ConditionalFormatting\ConditionalFormatting.cs" />
+    <Compile Include="ConditionalFormatting\ConditionalFormattingTest.cs" />
     <Compile Include="Drawing\Slicers\ExcelSlicerDrawingTest.cs" />
     <Compile Include="Drawing\Slicers\OlapDataNodeTest.cs" />
     <Compile Include="Drawing\Slicers\PivotTableNodeTest.cs" />
